Revert "update: organize component imports by indexing"
This reverts commit c4ff5bfd61cd38408959faada3d82069c62da2c7.
