+++ src/components/App/App.jsx
@@ -1,4 +1,7 @@
-import { ScrollTop, MainHeader, MainContent, MainFooter } from "../../components";
+import { MainContent } from "../MainContent/MainContent.jsx";
+import { MainFooter } from "../MainFooter/MainFooter";
+import { MainHeader } from "../MainHeader/MainHeader.jsx";
+import {ScrollTop} from "../ScrollTop/ScrollTop";
 
 import "./App.css";
