--- src/components/MainContent/About/index.js
@@ -1,1 +1,0 @@
-export * from './About.jsx'
