--- src/components/MainContent/Contact/index.js
@@ -1,1 +1,0 @@
-export * from './Contact.jsx'
