--- src/components/MainContent/Hero/index.js
@@ -1,1 +1,0 @@
-export * from './Hero.jsx'
