--- src/components/MainContent/Projects/ProjectCard/index.js
@@ -1,1 +1,0 @@
-export * from './ProjectCard.jsx'
