--- src/components/MainContent/Projects/ProjectFilter/index.js
@@ -1,1 +1,0 @@
-export * from './ProjectFilter.jsx'
