--- src/components/MainContent/Projects/ProjectModal/index.js
@@ -1,1 +1,0 @@
-export * from './ProjectModal.jsx'
