--- src/components/MainContent/Projects/index.js
@@ -1,1 +1,0 @@
-export * from './Projects.jsx'
