--- src/components/MainContent/Skills/SkillCard/index.js
@@ -1,1 +1,0 @@
-export * from './SkillCard.jsx'
