--- src/components/MainContent/Skills/SkillModal/index.js
@@ -1,1 +1,0 @@
-export * from './SkillModal.jsx'
