--- src/components/MainContent/Skills/index.js
@@ -1,1 +1,0 @@
-export * from './Skills.jsx'
