--- src/components/MainContent/index.js
@@ -1,1 +1,0 @@
-export * from './MainContent.jsx'
