--- src/components/MainFooter/index.js
@@ -1,1 +1,0 @@
-export * from './MainFooter.jsx'
