--- src/components/MainHeader/HeaderMenu/index.js
@@ -1,1 +1,0 @@
-export * from './HeaderMenu.jsx'
