+++ src/components/MainHeader/MainHeader.jsx
@@ -1,4 +1,4 @@
-import { HeaderMenu } from "./HeaderMenu";
+import { HeaderMenu } from "./HeaderMenu/HeaderMenu";
 
 import "./MainHeader.css";
