--- src/components/MainHeader/index.js
@@ -1,1 +1,0 @@
-export * from './MainHeader.jsx'
