--- src/components/Modal/index.js
@@ -1,1 +1,0 @@
-export * from './Modal.jsx'
