--- src/components/Navbar/index.js
@@ -1,1 +1,0 @@
-export * from './Navbar.jsx'
