--- src/components/ScrollTop/index.js
@@ -1,1 +1,0 @@
-export * from './ScrollTop.jsx'
