--- src/components/index.js
@@ -1,18 +1,0 @@
-export * from './App'
-export * from './MainFooter'
-export * from './MainContent'
-export * from './MainContent/Hero'
-export * from './MainContent/About'
-export * from './MainContent/Projects'
-export * from './MainContent/Projects/ProjectCard'
-export * from './MainContent/Projects/ProjectFilter'
-export * from './MainContent/Projects/ProjectModal'
-export * from './MainContent/Skills'
-export * from './MainContent/Skills/SkillCard'
-export * from './MainContent/Skills/SkillModal'
-export * from './MainContent/Contact'
-export * from './MainHeader'
-export * from './MainHeader/HeaderMenu'
-export * from './ScrollTop'
-export * from './Modal'
-export * from './Navbar'
